test(map): move to `map`.
